backend/requirements.txt
asgiref==3.7.2
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
Pillow==9.5.0
sqlparse==0.3.1
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv
typing_extensions==4.6.3
tzdata==2023.3
